Refactored Media Structure
There is now a development and production branch for the static media files.  The development branch contains the raw, commented files, and the production branch contains compressed versions.  Which branch is used is determined by the value of WP_DEBUG.  The build script was also updated to place the files in their proper locations.